refactor(scrapers.tasks.update_from_text):  reuse make_citation in update_from_text
Solves #4903

- Move make_citation from cl_scrape_opinions into scrapers.utils
- Move citation_is_duplicated from cl_back_scrape_citations into scrapers.utils
- Delete scraped_citation_object_is_valid, now we rely on eyecite, used by make_citation
- Refactor test site to account for changes
