+++ cl/scrapers/tests.py
@@ -40,7 +40,6 @@ from cl.scrapers.utils import (
     get_binary_content,
     get_existing_docket,
     get_extension,
-    scraped_citation_object_is_valid,
     update_or_create_docket,
 )
 from cl.search.factories import (
@@ -992,26 +991,6 @@ class UpdateFromTextCommandTest(TestCase):
             self.opinion_2020_unpub.cluster.docket.docket_number,
             "13",
             "Unpublished docket should not be modified",
-        )
-
-    def test_scraped_citation_object_is_valid(self):
-        """Can we validate Citation dicts got from `Site.extract_from_text`"""
-        bad_type = {"reporter": "WI", "type": Citation.FEDERAL}
-        self.assertFalse(
-            scraped_citation_object_is_valid(bad_type),
-            "Citation should be marked as invalid. Type does not match reporter",
-        )
-
-        bad_reporter = {"reporter": "Some text"}
-        self.assertFalse(
-            scraped_citation_object_is_valid(bad_reporter),
-            "Citation should be marked as invalid. Reporter does not exist",
-        )
-
-        valid_citation = {"reporter": "WI", "type": Citation.NEUTRAL}
-        self.assertTrue(
-            scraped_citation_object_is_valid(valid_citation),
-            "Citation object should be marked as valid",
         )
